+++ src/Tachycardia.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Nexus\PHPUnit\Extension;
+namespace Nexus\PHPUnit\Tachycardia;
 
-use Nexus\PHPUnit\Extension\Util\GithubMonitor;
-use Nexus\PHPUnit\Extension\Util\Parser;
-use Nexus\PHPUnit\Extension\Util\TestCase;
-use Nexus\PHPUnit\Extension\Util\TimeState;
+use Nexus\PHPUnit\Tachycardia\Util\GithubMonitor;
+use Nexus\PHPUnit\Tachycardia\Util\Parser;
+use Nexus\PHPUnit\Tachycardia\Util\TestCase;
+use Nexus\PHPUnit\Tachycardia\Util\TimeState;
 use PHPUnit\Runner\AfterLastTestHook;
 use PHPUnit\Runner\AfterSuccessfulTestHook;
 use PHPUnit\Runner\BeforeFirstTestHook;